Fixed different ad versions showing
