+++ src/app/actions.ts
@@ -3,7 +3,7 @@
 import { db } from "@/db";
 import { business_pitches, business_accounts } from "@/db/schema";
 import { fetchFeaturedMedia } from "@/lib/s3_utils";
-import { gt, desc, inArray, max, and, sql, eq } from "drizzle-orm";
+import { gt, desc, or, max, and, sql, eq } from "drizzle-orm";
 
 const PitchAd = {
     id: business_pitches.pitch_id,
@@ -21,20 +21,20 @@ const PitchAd = {
  */
 export async function getAdvertisementPitches() {
     // group by pitch_id to get the latest version of each pitch
-    const latestPitches = await db
-        .select({
-            pitch_id: business_pitches.pitch_id,
-            maxVersion: max(business_pitches.version),
-        })
-        .from(business_pitches)
-        .groupBy(business_pitches.pitch_id);
-
+    const latestPitches = await getAllLatestPitches();
+    
+    const conditions = latestPitches.map(lp =>
+        and(
+            eq(business_pitches.pitch_id, lp.pitch_id),
+            eq(business_pitches.version, lp.version)
+        )
+    );
     // get the full details of these latest pitches
     const pitches = await db.select(PitchAd).from(business_pitches)
         .where(
             and(
                 gt(business_pitches.adverts_available, 0),
-                inArray(business_pitches.pitch_id, latestPitches.map(p => p.pitch_id))
+                or(...conditions)
             )
         )
         .orderBy(desc(business_pitches.adverts_available));
@@ -56,9 +56,51 @@ export async function getAdvertisementPitches() {
 export async function updateAdvertCount(pitchInstanceId: string) {
     await db
         .update(business_pitches)
-        .set({ 
+        .set({
             adverts_available: sql`${business_pitches.adverts_available} - 1`,
             total_advert_clicks: sql`${business_pitches.total_advert_clicks} + 1`
-         })
+        })
         .where(eq(business_pitches.instance_id, pitchInstanceId));
+}
+
+/**
+ * Get the latest version of a specific pitch by pitch ID.
+ * @param pitchId Pitch ID to fetch the latest version for
+ * @returns The latest pitch version or undefined if not found
+ */
+export async function getLatestPitchVersion(pitchId: string) {
+    const [latestPitch] = await db
+        .select()
+        .from(business_pitches)
+        .where(eq(business_pitches.pitch_id, pitchId))
+        .orderBy(desc(business_pitches.version))
+        .limit(1);
+    return latestPitch;
+}
+
+/**
+ * Get the latest version of all pitches.
+ * @returns All latest pitch versions
+ */
+async function getAllLatestPitches() {
+    const latestVersions = await db
+        .select({
+            pitch_id: business_pitches.pitch_id,
+            max_version: max(business_pitches.version).as("max_version"),
+        })
+        .from(business_pitches)
+        .groupBy(business_pitches.pitch_id)
+        .as("latest_versions");
+ 
+    const pitches = await db
+        .select()
+        .from(business_pitches)
+        .innerJoin(
+            latestVersions,
+            and(
+                eq(business_pitches.pitch_id, latestVersions.pitch_id),
+                eq(business_pitches.version, latestVersions.max_version),
+            )
+        );
+    return pitches.map(p => p.business_pitches);
 }
